refactor(analysis-types): extract halstead DTO
